Adding Pidove; updating Cheat codes and Action Replay

+++ catscripts/Entertainment/Video_games/Cheat_codes/CheatCodes.sh
@@ -1,29 +1,39 @@
 #!/bin/bash
 
-egrep -i 'cheat code' newpages.txt >> CheatCodes.txt
-egrep -i 'cheats for' newpages.txt >> CheatCodes.txt
+KEYWORDS_CHEATCODES="cheat(| )code|cheats(| )for"
+KEYWORDS_ACTIONREPLAY="Action(| )Replay|\bar(| )code"
+KEYWORDS_CHEATCODES_EXCLUDE="$KEYWORDS_ACTIONREPLAY"
 
-egrep -i 'Action Replay' newpages.txt >> ActionReplay.txt
-egrep -i '\bar codes' newpages.txt >> ActionReplay.txt 
-
-
-CHEATCODES=`stat --print=%s CheatCodes.txt`
-ACTIONREPLAY=`stat --print=%s ActionReplay.txt`
-
-
-if [ $CHEATCODES -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="CheatCodes.txt"
-  export CATNAME="Cheat codes"
-  $CATEGORIZE
-fi
+  
+  debug_start "Cheat codes"
 
-if [ $ACTIONREPLAY -ne 0 ];
-then
-  export CATFILE="ActionReplay.txt"
-  export CATNAME="Action Replay"
-  $CATEGORIZE
-fi
 
-rm CheatCodes.txt
-rm ActionReplay.txt
+  CHEATCODES="$(egrep -i "$KEYWORDS_CHEATCODES" newpages.txt | egrep -iv "$KEYWORDS_CHEATCODES_EXCLUDE")"
+  ACTIONREPLAY="$(egrep -i "$KEYWORDS_ACTIONREPLAY" newpages.txt)"
+
+
+  if [ "$CHEATCODES" != "" ];
+  then
+    printf "%s" "$CHEATCODES" > CheatCodes.txt
+    export CATFILE="CheatCodes.txt"
+    export CATNAME="Cheat codes"
+    $CATEGORIZE
+    rm CheatCodes.txt
+    unset CHEATCODES
+  fi
+
+  if [ "$ACTIONREPLAY" != "" ];
+  then
+    printf "%s" "$ACTIONREPLAY" > ActionReplay.txt
+    export CATFILE="ActionReplay.txt"
+    export CATNAME="Action Replay"
+    $CATEGORIZE
+    rm ActionReplay.txt
+    unset ACTIONREPLAY
+  fi
+
+  debug_end "Cheat codes"
+
+fi
